requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-groq
langchain-google-genai
langchain-mistralai
langchain_openrouter
langchain-tavily
python-dotenv
openai
groq
google-generativeai
streamlit
tavily-python
beautifulsoup4

